Exception handler (#2)
* type of dto and entity fields were changed

* response status was changed for POST

* type of dto and entity fields were changed

* dto and entity fields were changed

* new dto`s for reference entities

* new dto`s for reference entities

* new dto for exceptions

* handling single exceptions was modified

* new exceptions dto

* handling multiple exceptions was modified

+++ src/main/java/itacad/aliaksandrkryvapust/myfitapp/core/dto/input/ProductDtoInput.java
@@ -21,14 +21,14 @@ public class ProductDtoInput {
     private final Integer calories;
     @NotNull(message = "proteins cannot be null")
     @DecimalMin(value = "0", message = "proteins should be zero or greater")
-    private final Integer proteins;
+    private final Double proteins;
     @NotNull(message = "fats cannot be null")
     @DecimalMin(value = "0", message = "fats should  be zero or greater")
-    private final Integer fats;
+    private final Double fats;
     @NotNull(message = "carbohydrates cannot be null")
     @DecimalMin(value = "0", message = "carbohydrates should  be zero or greater")
-    private final Integer carbohydrates;
+    private final Double carbohydrates;
     @NotNull(message = "weight cannot be null")
     @Positive(message = "weight should be positive")
-    private final Double weight;
+    private final Integer weight;
 }
